+++ proof_of_concepts/dash_app/app.py
@@ -10,7 +10,7 @@ from typing import Generator, Any
 
 import networkx as nx
 
-from tel2puml.pv_to_puml import (
+from tel2puml.pv_to_puml.pv_to_puml import (
     pv_jobs_from_folder_to_puml_file,
     pv_jobs_from_folder_to_event_sequence_streams,
 )
