+++ tel2puml/__main__.py
@@ -19,7 +19,7 @@ Example:
 """
 
 import argparse
-from tel2puml.pv_to_puml import (
+from tel2puml.pv_to_puml.pv_to_puml import (
     pv_jobs_from_folder_to_puml_file,
     pv_jobs_from_files_to_puml_file,
     pv_events_from_folder_to_puml_file,
